Add server-side pagination and filtering to admin tabs
Refactored admin tabs for authors, genders, and editorials to use server-side pagination and filtering via updated TRPC endpoints. Updated the catalog router to support search, pagination, and total count for these entities. Simplified the HomeAdminContent component to delegate tab content and removed redundant local filtering and pagination logic. Improved the books admin tab with advanced filter controls and search bar.
