+++ src/app/_components/home/content/_components/admin-editorials-tab.tsx
@@ -1,4 +1,4 @@
-import { useState, useMemo, useEffect } from "react";
+import { useState, useEffect } from "react";
 import { api } from "~/trpc/react";
 import { Button } from "~/components/ui/button";
 import {
@@ -32,19 +32,18 @@ export function AdminEditorialsTab() {
   const [deleteEditorial, setDeleteEditorial] = useState<Editorial | null>(null);
   const utils = api.useUtils();
 
-  const { data, isLoading } = api.catalog.getAllEditorials.useQuery();
-  const allEditorials = data?.response ?? [];
-  // Filtrado frontend
-  const filtered = search.trim()
-    ? allEditorials.filter(e => e.name && e.name.toLowerCase().includes(search.trim().toLowerCase()))
-    : allEditorials;
-  // Paginación frontend
-  const totalPages = Math.max(1, Math.ceil(filtered.length / pageSize));
-  const editorials = filtered.slice((page - 1) * pageSize, page * pageSize);
-  useEffect(() => { setPage(1); }, [search]);
+  const { data, isLoading } = api.catalog.getAllEditorials.useQuery({
+    search: search.trim() || undefined,
+    page,
+    limit: pageSize,
+  });
+
+  const editorials = data?.response ?? [];
+  const pagination = data?.pagination;
+
   useEffect(() => {
-    if (page > totalPages) setPage(totalPages);
-  }, [totalPages, page]);
+    setPage(1);
+  }, [search]);
 
   // Mutations
   const createEditorial = api.catalog.createEditorial.useMutation({
@@ -150,14 +149,16 @@ export function AdminEditorialsTab() {
           </div>
         </CardContent>
       </Card>
-      {totalPages > 1 && (
+      {pagination && pagination.totalPages > 1 && (
         <PaginationControls
           className="mt-6"
-          currentPage={page}
-          totalPages={totalPages}
-          hasNextPage={page < totalPages}
-          hasPreviousPage={page > 1}
-          onPageChange={p => { if (p >= 1 && p <= totalPages) setPage(p); }}
+          currentPage={pagination.page}
+          totalPages={pagination.totalPages}
+          hasNextPage={pagination.hasNextPage}
+          hasPreviousPage={pagination.hasPreviousPage}
+          onPageChange={p => {
+            if (p >= 1 && p <= pagination.totalPages) setPage(p);
+          }}
         />
       )}
       {/* Modal de agregar/editar */}
