+++ src/server/api/routers/catalog.ts
@@ -1,27 +1,116 @@
 import { createTRPCRouter, publicProcedure } from "../trpc";
 import { authors, editorials, genders, locations } from "~/server/db/schemas";
 import { z } from "zod";
-import { eq } from "drizzle-orm";
+import { eq, or, ilike, count } from "drizzle-orm";
 
 export const catalogRouter = createTRPCRouter({
-  getAllAuthors: publicProcedure.query(async ({ ctx }) => {
-    const allAuthors = await ctx.db.select().from(authors);
+  getAllAuthors: publicProcedure
+    .input(
+      z
+        .object({
+          search: z.string().optional(),
+          page: z.number().min(1).default(1),
+          limit: z.number().min(1).max(100).default(10),
+        })
+        .optional(),
+    )
+    .query(async ({ ctx, input }) => {
+      const { search, page = 1, limit = 10 } = input ?? {};
+      const offset = (page - 1) * limit;
 
-    return {
-      success: true,
-      method: "GET",
-      response: allAuthors,
-    };
-  }),
-  getAllGenders: publicProcedure.query(async ({ ctx }) => {
-    const allGenders = await ctx.db.select().from(genders);
+      // Build where condition
+      const whereCondition = search
+        ? or(
+            ilike(authors.name, `%${search}%`),
+            ilike(authors.middleName, `%${search}%`),
+            ilike(authors.lastName, `%${search}%`),
+          )
+        : undefined;
 
-    return {
-      success: true,
-      method: "GET",
-      response: allGenders,
-    };
-  }),
+      // Get total count
+      const totalCountResult = await ctx.db
+        .select({ count: count() })
+        .from(authors)
+        .where(whereCondition);
+      const totalCount = totalCountResult[0]?.count ?? 0;
+
+      // Get paginated results
+      const allAuthors = await ctx.db
+        .select()
+        .from(authors)
+        .where(whereCondition)
+        .limit(limit)
+        .offset(offset)
+        .orderBy(authors.lastName, authors.name);
+
+      const totalPages = Math.ceil(totalCount / limit);
+
+      return {
+        success: true,
+        method: "GET",
+        response: allAuthors,
+        pagination: {
+          page,
+          limit,
+          totalCount,
+          totalPages,
+          hasNextPage: page < totalPages,
+          hasPreviousPage: page > 1,
+        },
+      };
+    }),
+
+  getAllGenders: publicProcedure
+    .input(
+      z
+        .object({
+          search: z.string().optional(),
+          page: z.number().min(1).default(1),
+          limit: z.number().min(1).max(100).default(10),
+        })
+        .optional(),
+    )
+    .query(async ({ ctx, input }) => {
+      const { search, page = 1, limit = 10 } = input ?? {};
+      const offset = (page - 1) * limit;
+
+      // Build where condition
+      const whereCondition = search
+        ? ilike(genders.name, `%${search}%`)
+        : undefined;
+
+      // Get total count
+      const totalCountResult = await ctx.db
+        .select({ count: count() })
+        .from(genders)
+        .where(whereCondition);
+      const totalCount = totalCountResult[0]?.count ?? 0;
+
+      // Get paginated results
+      const allGenders = await ctx.db
+        .select()
+        .from(genders)
+        .where(whereCondition)
+        .limit(limit)
+        .offset(offset)
+        .orderBy(genders.name);
+
+      const totalPages = Math.ceil(totalCount / limit);
+
+      return {
+        success: true,
+        method: "GET",
+        response: allGenders,
+        pagination: {
+          page,
+          limit,
+          totalCount,
+          totalPages,
+          hasNextPage: page < totalPages,
+          hasPreviousPage: page > 1,
+        },
+      };
+    }),
 
   createAuthor: publicProcedure
     .input(
@@ -131,15 +220,57 @@ export const catalogRouter = createTRPCRouter({
       };
     }),
 
-  getAllEditorials: publicProcedure.query(async ({ ctx }) => {
-    const allEditorials = await ctx.db.select().from(editorials);
+  getAllEditorials: publicProcedure
+    .input(
+      z
+        .object({
+          search: z.string().optional(),
+          page: z.number().min(1).default(1),
+          limit: z.number().min(1).max(100).default(10),
+        })
+        .optional(),
+    )
+    .query(async ({ ctx, input }) => {
+      const { search, page = 1, limit = 10 } = input ?? {};
+      const offset = (page - 1) * limit;
 
-    return {
-      success: true,
-      method: "GET",
-      response: allEditorials,
-    };
-  }),
+      // Build where condition
+      const whereCondition = search
+        ? ilike(editorials.name, `%${search}%`)
+        : undefined;
+
+      // Get total count
+      const totalCountResult = await ctx.db
+        .select({ count: count() })
+        .from(editorials)
+        .where(whereCondition);
+      const totalCount = totalCountResult[0]?.count ?? 0;
+
+      // Get paginated results
+      const allEditorials = await ctx.db
+        .select()
+        .from(editorials)
+        .where(whereCondition)
+        .limit(limit)
+        .offset(offset)
+        .orderBy(editorials.name);
+
+      const totalPages = Math.ceil(totalCount / limit);
+
+      return {
+        success: true,
+        method: "GET",
+        response: allEditorials,
+        pagination: {
+          page,
+          limit,
+          totalCount,
+          totalPages,
+          hasNextPage: page < totalPages,
+          hasPreviousPage: page > 1,
+        },
+      };
+    }),
 
   createEditorial: publicProcedure
     .input(
